Refactor new tests to reduce DB-chatter
