+++ tests/Filter.NPoco.Tests/SqlBuilderExtensionsTests.cs
@@ -1,105 +1,34 @@
-﻿using NPoco;
-using RimDev.Automation.Sql;
-using RimDev.Filter.NPoco;
+﻿using RimDev.Filter.NPoco;
+using RimDev.Filter.Range;
 using RimDev.Filter.Range.Generic;
 using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
 using System.Linq;
 using Xunit;
 
-namespace RimDev.Filter.Tests
+namespace Filter.NPoco.Tests
 {
     public class SqlBuilderExtensionsTests
     {
-        public class Filter : SqlBuilderExtensionsTests, IDisposable
+        public abstract class Filter : SqlBuilderExtensionsTests, IClassFixture<DatabaseFixture>
         {
-            public Filter()
+            public Filter(DatabaseFixture fixture)
             {
-                localDb = new LocalDb(version: "v13.0");
-
-                using (var connection = new SqlConnection(localDb.ConnectionString))
-                {
-                    connection.Open();
-
-                    using (var command = new SqlCommand())
-                    {
-                        command.Connection = connection;
-                        command.CommandText = @"
-create table Person(
-    Id int identity not null,
-    FavoriteDate datetime,
-    FavoriteDateTimeOffset datetimeoffset,
-    FavoriteLetter nchar(1),
-    FavoriteNumber int,
-    FirstName nvarchar(50),
-    LastName nvarchar(50),
-    Rating decimal(5,2) null
-)";
-
-                        command.ExecuteNonQuery();
-                    }
-                }
-
-                Database = new Database(localDb.ConnectionString, DatabaseType.SqlServer2008);
-
-                Database.InsertBulk(people);
+                Fixture = fixture;
             }
 
-            [TableName("Person")]
-            [PrimaryKey(nameof(Id))]
-            public class Person
-            {
-                public int Id { get; set; }
-                public DateTime FavoriteDate { get; set; }
-                public DateTimeOffset FavoriteDateTimeOffset { get; set; }
-                public char FavoriteLetter { get; set; }
-                public int FavoriteNumber { get; set; }
-                public string FirstName { get; set; }
-                public string LastName { get; set; }
-                public decimal? Rating { get; set; }
-            }
-
-            protected readonly IDatabase Database;
-
-            private readonly IEnumerable<Person> people = new List<Person>()
-            {
-                new Person()
-                {
-                    FavoriteDate = DateTime.Parse("2000-01-01"),
-                    FavoriteDateTimeOffset = DateTimeOffset.Parse("2010-01-01"),
-                    FavoriteLetter = 'a',
-                    FavoriteNumber = 5,
-                    FirstName = "John",
-                    LastName = "Doe"
-                },
-                new Person()
-                {
-                    FavoriteDate = DateTime.Parse("2000-01-02"),
-                    FavoriteDateTimeOffset = DateTimeOffset.Parse("2010-01-02"),
-                    FavoriteLetter = 'b',
-                    FavoriteNumber = 10,
-                    FirstName = "Tim",
-                    LastName = "Smith",
-                    Rating = 4.5m
-                },
-            };
-
-            private LocalDb localDb;
-
-            public void Dispose()
-            {
-                localDb.Dispose();
-                Database.Dispose();
-            }
+            protected readonly DatabaseFixture Fixture;
         }
 
         public class ConstantFilters : Filter
         {
+            public ConstantFilters(DatabaseFixture fixture)
+                : base(fixture)
+            { }
+
             [Fact]
             public void Should_filter_when_property_types_match_as_constant_string()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -114,7 +43,7 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_constant_integer()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -129,7 +58,7 @@ create table Person(
             [Fact]
             public void Should_filter_multiple_properties()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -144,7 +73,7 @@ create table Person(
             [Fact]
             public void Should_not_throw_if_filter_does_not_contain_valid_properties()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -158,7 +87,7 @@ create table Person(
             [Fact]
             public void Should_not_throw_if_filter_constant_type_does_not_match()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -172,10 +101,14 @@ create table Person(
 
         public class EnumerableFilters : Filter
         {
+            public EnumerableFilters(DatabaseFixture fixture)
+                : base(fixture)
+            { }
+
             [Fact]
             public void Should_bypass_filter_when_empty_collection()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -189,7 +122,7 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_enumerable_string()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -204,7 +137,7 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_enumerable_integer()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -219,7 +152,7 @@ create table Person(
             [Fact]
             public void Should_be_able_to_handle_nullable_destination_and_primitive_filter()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -232,7 +165,7 @@ create table Person(
             [Fact]
             public void Should_be_able_to_handle_nullable_destination_and_nullable_primitive_filter_as_collection()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -245,7 +178,7 @@ create table Person(
             [Fact]
             public void Should_be_able_to_handle_nullable_destination_and_nullable_primitive_filter()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -258,19 +191,11 @@ create table Person(
             [Fact]
             public void Should_be_able_to_handle_nullable_source()
             {
-                var people = Database.Query<Person>().ToList();
-                people.ForEach(x => x.Rating = null);
-
-                foreach (var person in people)
-                {
-                    Database.Update(person);
-                }
-
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        Rating = new decimal?[] { 4.5m }
+                        DONOTUSE = new string[] { "whatever" }
                     });
 
                 Assert.Empty(@return);
@@ -279,16 +204,20 @@ create table Person(
 
         public class RangeFilters : Filter
         {
+            public RangeFilters(DatabaseFixture fixture)
+                : base(fixture)
+            { }
+
             [Theory,
             InlineData("(,5]"),
             InlineData("(-∞,5]")]
             public void Should_filter_open_ended_lower_bound(string value)
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<int>(value)
+                        FavoriteNumber = Range.FromString<int>(value)
                     });
 
                 Assert.NotNull(@return);
@@ -301,11 +230,11 @@ create table Person(
             InlineData("(5,+∞)")]
             public void Should_filter_open_ended_upper_bound(string value)
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<int>(value)
+                        FavoriteNumber = Range.FromString<int>(value)
                     });
 
                 Assert.NotNull(@return);
@@ -316,7 +245,7 @@ create table Person(
             [Fact]
             public void Should_filter_for_concrete_range()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
@@ -337,11 +266,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_byte()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<byte>("[5,5]")
+                        FavoriteNumber = Range.FromString<byte>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -352,11 +281,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_char()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteLetter = Range.Range.FromString<char>("[a,b)")
+                        FavoriteLetter = Range.FromString<char>("[a,b)")
                     });
 
                 Assert.NotNull(@return);
@@ -367,11 +296,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_datetime()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteDate = Range.Range.FromString<DateTime>("[2000-01-01,2000-01-02)")
+                        FavoriteDate = Range.FromString<DateTime>("[2000-01-01,2000-01-02)")
                     });
 
                 Assert.NotNull(@return);
@@ -382,11 +311,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_datetimeoffset()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteDateTimeOffset = Range.Range.FromString<DateTimeOffset>("[2010-01-01,2010-01-02)")
+                        FavoriteDateTimeOffset = Range.FromString<DateTimeOffset>("[2010-01-01,2010-01-02)")
                     });
 
                 Assert.NotNull(@return);
@@ -397,11 +326,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_decimal()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<decimal>("[4.5,5]")
+                        FavoriteNumber = Range.FromString<decimal>("[4.5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -412,11 +341,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_double()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<double>("[4.5,5]")
+                        FavoriteNumber = Range.FromString<double>("[4.5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -427,11 +356,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_float()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<float>("[4.5,5]")
+                        FavoriteNumber = Range.FromString<float>("[4.5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -442,11 +371,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_int()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<int>("[5,5]")
+                        FavoriteNumber = Range.FromString<int>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -457,11 +386,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_long()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<long>("[5,5]")
+                        FavoriteNumber = Range.FromString<long>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -472,11 +401,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_sbyte()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<sbyte>("[5,5]")
+                        FavoriteNumber = Range.FromString<sbyte>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -487,11 +416,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_short()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<short>("[5,5]")
+                        FavoriteNumber = Range.FromString<short>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -502,11 +431,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_uint()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<uint>("[5,5]")
+                        FavoriteNumber = Range.FromString<uint>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -517,11 +446,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_ulong()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<ulong>("[5,5]")
+                        FavoriteNumber = Range.FromString<ulong>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -532,11 +461,11 @@ create table Person(
             [Fact]
             public void Should_filter_when_property_types_match_as_range_ushort()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
-                        FavoriteNumber = Range.Range.FromString<ushort>("[5,5]")
+                        FavoriteNumber = Range.FromString<ushort>("[5,5]")
                     });
 
                 Assert.NotNull(@return);
@@ -547,7 +476,7 @@ create table Person(
             [Fact]
             public void Should_be_able_to_handle_nullable_source()
             {
-                var @return = Database
+                var @return = Fixture.Database
                     .With<Person>("where /**where**/")
                     .Fetch(new
                     {
